add merge gh workflow, refactor location of problems
