+++ .github/workflows/build-test-merge.yaml
@@ -1,0 +1,26 @@
+name: Build and Run Unit Tests on merge
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.x'
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+
+      - name: Test with unittest
+        run: |
+          python -m unittest discover -s tests -p '*_test.py'
