--- tests/algorithm/largest_palindrome_product_test.py
@@ -1,10 +1,0 @@
-import unittest
-
-from app.algorithm.largest_palindrome_product import LargestPalindromeProduct
-
-
-class LargestPalindromeProductTest(unittest.TestCase):
-
-    def test_brute_force(self):
-        algo = LargestPalindromeProduct()
-        self.assertEqual(906609, algo.brute_force())
